requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psycopg2==2.9.3
psycopg2-binary==2.9.4
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.2.1
qrcode==7.3.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
